% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ggpgam.R
\name{pred_gam}
\alias{pred_gam}
\title{Predict from a fitted GGP-GAM model}
\usage{
pred_gam(fit, X_new, coords_new)
}
\arguments{
\item{fit}{A fitted model returned by [fit_gam()].}

\item{X_new}{Numeric matrix of new covariates.}

\item{coords_new}{Numeric matrix of new coordinates.}
}
\value{
Numeric vector of predicted responses.
}
\description{
Predict from a fitted GGP-GAM model
}
