% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_mgwr.R
\name{pred_mgwr}
\alias{pred_mgwr}
\title{Predict from a fitted MGWR model}
\usage{
pred_mgwr(model, X_new, coords_new)
}
\arguments{
\item{model}{A fitted model returned by [fit_mgwr()].}

\item{X_new}{Numeric matrix of new covariates.}

\item{coords_new}{Numeric matrix of new coordinates.}
}
\value{
Numeric vector of predicted responses.
}
\description{
Predict from a fitted MGWR model
}
